requirements.txt
streamlit
opencv-python
pandas
numpy
scikit-learn
Pillow
